+++ packages/browser-extension/src/modules/ui/options.tsx
@@ -197,12 +197,6 @@ export const SetupForm: React.FC = () => {
               <td>Select tag</td>
               <td>Click on tag</td>
             </tr>
-            <tr>
-              <td>Re-focus search box</td>
-              <td>
-                <kbd className="key-name">/</kbd>
-              </td>
-            </tr>
           </tbody>
         </table>
       </section>