+++ packages/browser-extension/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.3.0
 
 - Added: Link to chrome shortcuts page
+- Changed: <kbd>Esc</kbd> to re-focus search box and close popup
 - Changed: Cursor style tweak
 - Chore: Removed unused package
 - Fixed: HTML talbe semantics
@@ -14,7 +15,7 @@
 - Added: In-app documentation
 - Changed: Improved performance and reduced bandwidth usage
 - Changed: Left click on link will close the popup
-- Changed: Default keyboard shortcut will be Alt+A
+- Changed: Default keyboard shortcut will be <kbd>Alt</kbd>+<kbd>A</kbd>
 - Fixed: Long tag caused overflow
 - Fixed: Typing into search box didn't reset scroll